refactor: update the users table.
Changes: id is now generated by the drizzle itself; email is now indexed in case of searching user by email.

Note: because the first migration contained 'id: serial', the old migration files were replaced with new ones in order to resolve an error that occurred when the type was changed to 'uuid'. This decision was made because it was the first and only migration.