refactor: add a hook which is a wrapper of useActionState hook.
The useActionState hook requires the action to accept the previous state as the first argument and the payload as the second. However, the previous state is never used in any of the current server actions.
Therefore, the wrapper (useServerAction) enables actions with a single argument (payload) to be used.
The returned action is also already wrapped in the startTransition function, meaning the returned function can be called without wrapping.